+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/BrowserMenuItem.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.menu
 
 import android.content.Context
 import android.view.View
+import mozilla.components.browser.menu.view.ExpandableLayout
 import mozilla.components.concept.menu.candidate.MenuCandidate
 
 /**
@@ -24,6 +25,13 @@ interface BrowserMenuItem {
      * item, like a tool bar, will have several.
      */
     val interactiveCount: () -> Int get() = { 1 }
+
+    /**
+     * Whether this menu item can serve as the limit of a collapsing menu.
+     *
+     * @see [ExpandableLayout]
+     */
+    val isCollapsingMenuLimit: Boolean get() = false
 
     /**
      * Returns the layout resource ID of the layout to be inflated for showing a menu item of this